Extract logic to handle the Bin value lookup.
