+++ src/components/Card/Card.tsx
@@ -4,13 +4,11 @@ import CardInput from './components/CardInput';
 import CoreProvider from '../../core/Context/CoreProvider';
 import getImage from '../../utils/get-image';
 import collectBrowserInfo from '../../utils/browserInfo';
-import fetchJSONData from '../../utils/fetch-json-data';
 import { CardElementData, CardElementProps } from './types';
-import { CbObjOnError } from '../internal/SecuredFields/lib/types';
+import handleBinLookUp from './handleBinLookUp';
 
 export class CardElement extends UIElement<CardElementProps> {
     public static type = 'scheme';
-    private currentRequestId;
 
     formatProps(props: CardElementProps) {
         return {
@@ -49,12 +47,12 @@ export class CardElement extends UIElement<CardElementProps> {
     }
 
     updateStyles(stylesObj) {
-        if (this.componentRef && this.componentRef.updateStyles) this.componentRef.updateStyles(stylesObj);
+        if (this.componentRef?.updateStyles) this.componentRef.updateStyles(stylesObj);
         return this;
     }
 
     setFocusOn(fieldName) {
-        if (this.componentRef && this.componentRef.setFocusOn) this.componentRef.setFocusOn(fieldName);
+        if (this.componentRef?.setFocusOn) this.componentRef.setFocusOn(fieldName);
         return this;
     }
 
@@ -64,80 +62,16 @@ export class CardElement extends UIElement<CardElementProps> {
     };
 
     processBinLookupResponse(binLookupObject) {
-        if (this.componentRef && this.componentRef.processBinLookupResponse) this.componentRef.processBinLookupResponse(binLookupObject);
+        if (this.componentRef?.processBinLookupResponse) this.componentRef.processBinLookupResponse(binLookupObject);
         return this;
     }
 
     handleUnsupportedCard(errObj) {
-        if (this.componentRef && this.componentRef.handleUnsupportedCard) this.componentRef.handleUnsupportedCard(errObj);
+        if (this.componentRef?.handleUnsupportedCard) this.componentRef.handleUnsupportedCard(errObj);
         return this;
     }
 
-    public onBinValue = callbackObj => {
-        // Allow way for merchant to disallow binLookup by specifically setting the prop to false
-        if (this.props.doBinLookup === false) {
-            if (this.props.onBinValue) this.props.onBinValue(callbackObj);
-            return;
-        }
-
-        // Do binLookup when encryptedBin property is present (and only if the merchant is using a clientKey)
-        if (callbackObj.encryptedBin && this.props.clientKey) {
-            // Store id of request we're about to make
-            this.currentRequestId = callbackObj.uuid;
-
-            fetchJSONData(
-                {
-                    path: `v1/bin/binLookup?token=${this.props.clientKey}`,
-                    loadingContext: this.props.loadingContext,
-                    method: 'POST',
-                    contentType: 'application/json'
-                },
-                {
-                    supportedBrands: this.props.brands,
-                    encryptedBin: callbackObj.encryptedBin,
-                    requestId: callbackObj.uuid // Pass id of request
-                }
-            ).then(data => {
-                // If response is the one we were waiting for...
-                if (data && data.requestId === this.currentRequestId) {
-                    // ...call processBinLookupResponse with the response object if it contains at least one supported brand
-                    if (data.supportedBrands && data.supportedBrands.length) {
-                        this.processBinLookupResponse(data);
-                        return;
-                    }
-                    // If we get here then no supported brands were found
-                    if (data.detectedBrands && data.detectedBrands.length) {
-                        const errObj: CbObjOnError = {
-                            type: 'card',
-                            fieldType: 'encryptedCardNumber',
-                            error: 'Unsupported card entered',
-                            binLookupBrands: data.detectedBrands
-                        };
-                        this.handleUnsupportedCard(errObj);
-                        return;
-                    }
-                    // A failed lookup will just contain requestId - we may still need to do something at this point
-                    // console.log('### Card::onBinValue:: binLookup response - no match found for request:', data.requestId);
-                }
-            });
-        } else if (this.currentRequestId) {
-            // If onBinValue callback is called AND we have been doing binLookup BUT passed object doesn't have an encryptedBin property
-            // - then the number of digits in number field has dropped below threshold for BIN lookup - so reset the UI
-            this.processBinLookupResponse(null);
-
-            this.currentRequestId = null; // Ignore any pending responses
-
-            // Reset any errors
-            const errObj: CbObjOnError = {
-                type: 'card',
-                fieldType: 'encryptedCardNumber',
-                error: ''
-            };
-            this.handleUnsupportedCard(errObj);
-        }
-
-        if (this.props.onBinValue) this.props.onBinValue(callbackObj);
-    };
+    public onBinValue = handleBinLookUp.bind(this);
 
     get isValid() {
         return !!this.state.isValid;
@@ -148,9 +82,10 @@ export class CardElement extends UIElement<CardElementProps> {
     }
 
     get brands(): { icon: any; name: string }[] {
-        if (this.props.brands) {
-            return this.props.brands.map(brand => ({
-                icon: getImage({ loadingContext: this.props.loadingContext })(brand),
+        const { brands, loadingContext } = this.props;
+        if (brands) {
+            return brands.map(brand => ({
+                icon: getImage({ loadingContext })(brand),
                 name: brand
             }));
         }
